Tower placement bug fixed (detection of other towers),Towers rotate to target in range, factored out tile and point conversion, started on bullet entity.

+++ core/src/Entity/EntityManager.java
@@ -45,7 +45,7 @@ public class EntityManager
     }
 
     /* Update method used to update all entities in the array list*/
-    public void updateEnts(int t) 
+    public void updateEnts(float t) 
     {
         for (Entity ents : e)
         {
